drm_hwcomposer: Make sure we check for new 16bit alpha value
In commit 025d0a65 ("Use 16 bit value for plane alpha"),
the alpha value was bumped to a 16bit value.

Unfortunately in drmdisplaycompositor.cpp we are still
comparing it against a 8bit 0xff value, which now fails
causing the display to not start.

So fix this comparing to the 16bit 0xffff value.

With this change the display starts up again.

Change-Id: Ic796b16fb57ae7cea58b5c66c456a58e3fb57ba5
Signed-off-by: John Stultz <john.stultz@linaro.org>